Move ShakeMapGrid and Point into grid module
