+++ sc/tests/impact.py
@@ -1,6 +1,6 @@
 import unittest
 from sc.app.impact import *
-from sc.app.objects import ShakeMapGrid, Point
+from sc.app.grid import ShakeMapGrid, Point
 from sc.app.orm import Aebm, Event, ShakeMap
 
 from .util import create_fac
